Split browser and browser context
