+++ lib/PuppeteerSharp/BrowserContext.cs
@@ -7,17 +7,8 @@ using PuppeteerSharp.Messaging;
 namespace PuppeteerSharp
 {
     /// <inheritdoc/>
-    public class BrowserContext : IBrowserContext
+    public abstract class BrowserContext : IBrowserContext
     {
-        private readonly Connection _connection;
-
-        internal BrowserContext(Connection connection, Browser browser, string contextId)
-        {
-            _connection = connection;
-            Browser = browser;
-            Id = contextId;
-        }
-
         /// <inheritdoc/>
         public event EventHandler<TargetChangedArgs> TargetChanged;
 
@@ -28,7 +19,7 @@ namespace PuppeteerSharp
         public event EventHandler<TargetChangedArgs> TargetDestroyed;
 
         /// <inheritdoc/>
-        public string Id { get; }
+        public string Id { get; protected init; }
 
         /// <inheritdoc/>
         public bool IsIncognito => Id != null;
@@ -37,57 +28,32 @@ namespace PuppeteerSharp
         public bool IsClosed => Browser.BrowserContexts().Contains(this);
 
         /// <inheritdoc cref="Browser"/>
-        public Browser Browser { get; }
+        public Browser Browser { get; protected init; }
 
         /// <inheritdoc/>
         IBrowser IBrowserContext.Browser => Browser;
 
         /// <inheritdoc/>
-        public ITarget[] Targets() => Array.FindAll(Browser.Targets(), target => target.BrowserContext == this);
+        public abstract ITarget[] Targets();
 
         /// <inheritdoc/>
         public Task<ITarget> WaitForTargetAsync(Func<ITarget, bool> predicate, WaitForOptions options = null)
             => Browser.WaitForTargetAsync((target) => target.BrowserContext == this && predicate(target), options);
 
-        /// <inheritdoc/>
-        public async Task<IPage[]> PagesAsync()
-        => (await Task.WhenAll(
-            Targets()
-                .Where(t =>
-                    t.Type == TargetType.Page ||
-                    (t.Type == TargetType.Other && Browser.IsPageTargetFunc(t as Target)))
-                .Select(t => t.PageAsync())).ConfigureAwait(false))
-            .Where(p => p != null).ToArray();
+        /// <inheritdoc />
+        public abstract Task<IPage[]> PagesAsync();
 
         /// <inheritdoc/>
-        public Task<IPage> NewPageAsync() => Browser.CreatePageInContextAsync(Id);
+        public abstract Task<IPage> NewPageAsync();
 
         /// <inheritdoc/>
-        public Task CloseAsync()
-        {
-            if (Id == null)
-            {
-                throw new PuppeteerException("Non-incognito profiles cannot be closed!");
-            }
-
-            return Browser.DisposeContextAsync(Id);
-        }
+        public abstract Task CloseAsync();
 
         /// <inheritdoc/>
-        public Task OverridePermissionsAsync(string origin, IEnumerable<OverridePermission> permissions)
-            => _connection.SendAsync("Browser.grantPermissions", new BrowserGrantPermissionsRequest
-            {
-                Origin = origin,
-                BrowserContextId = Id,
-                Permissions = permissions.ToArray(),
-            });
+        public abstract Task OverridePermissionsAsync(string origin, IEnumerable<OverridePermission> permissions);
 
         /// <inheritdoc/>
-        public Task ClearPermissionOverridesAsync()
-            => _connection.SendAsync("Browser.resetPermissions", new BrowserResetPermissionsRequest
-            {
-                BrowserContextId = Id,
-            });
+        public abstract Task ClearPermissionOverridesAsync();
 
         internal void OnTargetCreated(Browser browser, TargetChangedArgs args) => TargetCreated?.Invoke(browser, args);
